+++ provisioners/scripts/oracle/install_ol7_scala_sbt.sh
@@ -12,19 +12,21 @@ cd /usr/local/scala
 curdate=$(date +"%Y-%m-%d")
 
 # retrieve version number of latest release.
-curl --silent --dump-header curl-${sbthome}.${curdate}.out1 https://github.com/sbt/sbt/releases/latest --output /dev/null
-tr -d '\r' < curl-${sbthome}.${curdate}.out1 > curl-${sbthome}.${curdate}.out2
-sbtrelease=$(awk '/Location/ {print $2}' curl-${sbthome}.${curdate}.out2 | awk -F "/" '{print $8}')
+sbtrelease="v1.0.2"
+#curl --silent --dump-header curl-${sbthome}.${curdate}.out1 https://github.com/sbt/sbt/releases/latest --output /dev/null
+#tr -d '\r' < curl-${sbthome}.${curdate}.out1 > curl-${sbthome}.${curdate}.out2
+#sbtrelease=$(awk '/Location/ {print $2}' curl-${sbthome}.${curdate}.out2 | awk -F "/" '{print $8}')
 sbtdir="sbt"
 sbtfolder="${sbthome}-${sbtrelease:1}"
 sbtbinary="sbt-${sbtrelease:1}.tgz"
-rm -f curl-${sbthome}.${curdate}.out1
-rm -f curl-${sbthome}.${curdate}.out2
+#rm -f curl-${sbthome}.${curdate}.out1
+#rm -f curl-${sbthome}.${curdate}.out2
 
 # download sbt from cocl.us.
 wget --no-verbose https://cocl.us/${sbtbinary}
 
 # extract sbt binary.
+rm -f ${sbthome}
 tar -zxvf ${sbtbinary} --no-same-owner --no-overwrite-dir
 mv -f ${sbtdir} ${sbtfolder}
 chown -R root:root ./${sbtfolder}
@@ -56,8 +58,13 @@ sbt_env_name="SBT_HOME"
 sbt_env_value="/usr/local/scala/${sbthome}"
 
 cd /home/vagrant
-awk -v env_comment=${sbt_env_comment} -v env_name=${sbt_env_name} -v env_value=${sbt_env_value} -f /tmp/scripts/oracle/append_ol7_env_path.awk .bashrc > .bashrc.${curdate}.${sbthome}
-mv -f .bashrc.${curdate}.${sbthome} .bashrc
+
+# if env name exists (grep command), skip awk update.
+grep -qF "${sbt_env_name}" .bashrc || awk -v env_comment=${sbt_env_comment} -v env_name=${sbt_env_name} -v env_value=${sbt_env_value} -f /tmp/scripts/oracle/append_ol7_env_path.awk .bashrc > .bashrc.${curdate}.${sbthome}
+
+if [ -f ".bashrc.${curdate}.${sbthome}" ]; then
+  mv -f .bashrc.${curdate}.${sbthome} .bashrc
+fi
 
 chown vagrant:vagrant .bashrc
 chmod 644 .bashrc
